refactored code, added singleItempervenue property to request
