+++ src/assets/icons/calendar.tsx
@@ -1,14 +1,7 @@
 import * as React from "react";
 import Svg, { Path } from "react-native-svg";
-export const CalendarIcon = (props) => (
-  <Svg
-    width={14}
-    height={16}
-    viewBox="0 0 14 16"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
+export const CalendarIcon = () => (
+  <Svg width={14} height={16} viewBox="0 0 14 16" fill="none">
     <Path
       d="M12.5 2C13.3125 2 14 2.6875 14 3.5V14.5C14 15.3438 13.3125 16 12.5 16H1.5C0.65625 16 0 15.3438 0 14.5V3.5C0 2.6875 0.65625 2 1.5 2H3V0.375C3 0.1875 3.15625 0 3.375 0H4.625C4.8125 0 5 0.1875 5 0.375V2H9V0.375C9 0.1875 9.15625 0 9.375 0H10.625C10.8125 0 11 0.1875 11 0.375V2H12.5ZM12.3125 14.5C12.4062 14.5 12.5 14.4375 12.5 14.3125V5H1.5V14.3125C1.5 14.4375 1.5625 14.5 1.6875 14.5H12.3125Z"
       fill="#333659"
