+++ src/assets/icons/profile.tsx
@@ -1,14 +1,7 @@
 import * as React from "react";
 import Svg, { Path } from "react-native-svg";
-export const ProfileIcon = (props) => (
-  <Svg
-    width={24}
-    height={24}
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
+export const ProfileIcon = () => (
+  <Svg width={24} height={24} viewBox="0 0 24 24" fill="none">
     <Path
       d="M11.95 11.1C13.0241 11.1 14.0543 10.6733 14.8138 9.9138C15.5733 9.15428 16 8.12409 16 7.05C16 5.97591 15.5733 4.94572 14.8138 4.1862C14.0543 3.42669 13.0241 3 11.95 3C10.8759 3 9.84574 3.42669 9.08623 4.1862C8.32671 4.94572 7.90002 5.97591 7.90002 7.05C7.90002 8.12409 8.32671 9.15428 9.08623 9.9138C9.84574 10.6733 10.8759 11.1 11.95 11.1ZM11.95 4.15714C12.7173 4.15714 13.453 4.46193 13.9956 5.00447C14.5381 5.54701 14.8429 6.28275 14.8429 7.05C14.8429 7.81725 14.5381 8.553 13.9956 9.09553C13.453 9.63807 12.7173 9.94286 11.95 9.94286C11.1828 9.94286 10.447 9.63807 9.90449 9.09553C9.36196 8.55299 9.05717 7.81725 9.05717 7.05C9.05717 6.28275 9.36195 5.547 9.90449 5.00447C10.447 4.46193 11.1828 4.15714 11.95 4.15714Z"
       fill="#333659"
