+++ src/components/BottomSheet/BottomSheetFooterCustom.tsx
@@ -1,10 +1,11 @@
+import { FC } from "react";
 import { View, Text } from "react-native";
+
+import { PorstDetailsValue } from "../../types/responses/PostDetailsResponse";
+import { BottomSheetFooter } from "@gorhom/bottom-sheet";
 import { LikeIcon, MoreIcon, ShareIcon } from "../../assets/icons";
 
 import styles from "./styles";
-import { PorstDetailsValue } from "../../types/responses/PostDetailsResponse";
-import { FC } from "react";
-import { BottomSheetFooter } from "@gorhom/bottom-sheet";
 
 interface Props {
   vibeDetails: PorstDetailsValue | null;
