+++ src/components/DateSelectionModal/DateSelectionModal.tsx
@@ -3,16 +3,17 @@ import {
   Text,
   TouchableOpacity,
   View,
-  StyleSheet,
   Modal,
   SafeAreaView,
 } from "react-native";
 import { AntDesign } from "@expo/vector-icons";
 import DateTimePicker, { DateType } from "react-native-ui-datepicker";
-import { buttonsSelectData } from "../../utils/mockMarkers";
 import { formatDate } from "../../utils/helpersFunctions";
 import { StatusBar } from "expo-status-bar";
 import MyContext from "../../providers/mapContext/MapContext";
+
+import styles from "./styles";
+import { DATE_RANGES } from "../../constants/dateRanges";
 
 interface Props {
   onSelect: (date: string) => void;
@@ -48,9 +49,6 @@ export const DateSelectionModal: React.FC<Props> = ({
 
     setStartDate(new Date());
     setEndDate(new Date());
-
-    // setSelectedDate("Now");
-    // onCloseModal(false);
   };
 
   return (
@@ -73,7 +71,7 @@ export const DateSelectionModal: React.FC<Props> = ({
       </View>
 
       <View style={styles.modalContainer}>
-        {buttonsSelectData.map((button, index) => (
+        {DATE_RANGES.map((button, index) => (
           <TouchableOpacity
             key={index}
             style={styles.buttonDate}
@@ -150,14 +148,10 @@ export const DateSelectionModal: React.FC<Props> = ({
                   return;
                 }
                 setShowDatePiker(false);
-
-                // onSelect(`${preparedStartDate} - ${preparedEndDate}`);
                 setCustomDate({
                   startDate: startDate,
                   endDate: endDate,
                 });
-                // onSelect(`Custom`);
-                // onCloseModal(false);
               }}
             >
               <Text style={styles.textApplyDate}>Apply</Text>
@@ -169,86 +163,3 @@ export const DateSelectionModal: React.FC<Props> = ({
     </SafeAreaView>
   );
 };
-
-const styles = StyleSheet.create({
-  headerContainer: {
-    display: "flex",
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    borderBottomColor: "#D9DBEB80",
-    borderBottomWidth: 1,
-    paddingHorizontal: 20,
-    height: 70,
-  },
-  headerText: {
-    fontWeight: "600",
-    fontSize: 18,
-    color: "#333659",
-  },
-  headerCancel: {
-    fontWeight: "600",
-    fontSize: 16,
-    color: "#333659",
-  },
-  modalContainer: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    alignContent: "center",
-    height: "80%",
-    gap: 15,
-  },
-  buttonDate: {
-    position: "relative",
-    display: "flex",
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-    width: "80%",
-    borderBlockColor: "#D9DBEB",
-    borderWidth: 1,
-    borderRadius: 8,
-    height: 48,
-  },
-  textDate: {
-    fontSize: 16,
-    fontWeight: "600",
-    color: "#333659",
-  },
-  icon: {
-    position: "absolute",
-    top: 10,
-    left: 15,
-  },
-  containerPiker: {
-    flex: 1,
-    backgroundColor: "#F5FCFF",
-  },
-  containerDatePiker: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    height: "76%",
-  },
-  buttonApplyDate: {
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-    height: 50,
-    width: "80%",
-    borderRadius: 30,
-    backgroundColor: "#005DF2",
-    marginTop: 10,
-  },
-  textApplyDate: {
-    fontWeight: "600",
-    fontSize: 16,
-    color: "#fff",
-  },
-  dateTime: {
-    fontWeight: "600",
-    fontSize: 15,
-    color: "#005DF2",
-  },
-});
