+++ src/components/mapTopContainer/TotalResults.tsx
@@ -1,0 +1,31 @@
+import { View, Text } from "react-native";
+import styles from "./styles";
+import { EyeHiddenIcon, EyeIcon } from "../../assets/icons";
+import { FC } from "react";
+
+interface Props {
+  total: number;
+  visible: number;
+}
+
+export const TotalResults: FC<Props> = ({ total, visible }) => {
+  return total > 0 ? (
+    <View style={[styles.searchButton, styles.totalResults]}>
+      <View style={styles.resultContainer}>
+        <EyeHiddenIcon />
+        <Text style={styles.resultText}>
+          {total - visible >= 0 ? total - visible : "-"}
+        </Text>
+      </View>
+      <View style={styles.line} />
+      <View>
+        <EyeIcon />
+        <Text style={styles.resultText}>{visible}</Text>
+      </View>
+    </View>
+  ) : (
+    <View style={[styles.searchButton, styles.totalResults]}>
+      <Text>No results</Text>
+    </View>
+  );
+};
