+++ src/constants/dateRanges.ts
@@ -1,0 +1,11 @@
+export const DATE_RANGES = [
+  { label: "Now" },
+  { label: "Today" },
+  { label: "Tomorrow" },
+  { label: "Next 7 Days" },
+  { label: "Next Week" },
+  { label: "Next 14 Days" },
+  { label: "Next Month" },
+  { label: "Next 30 Days" },
+  { label: "Custom" },
+];
