+++ src/constants/mapStyleUrl.ts
@@ -1,3 +1,3 @@
-const mapboxStyleUrl = "mapbox://styles/vibespot/clvfm2nfq010401q14frq08bd";
+const MAPBOX_STYLE_URL = "mapbox://styles/vibespot/clvfm2nfq010401q14frq08bd";
 
-export default mapboxStyleUrl;
+export default MAPBOX_STYLE_URL;
