--- src/utils/mockMarkers.ts
@@ -1,31 +1,0 @@
-export const mockMarkers = [
-  {
-    id: "marker-1",
-    coordinate: [-73.989308, 40.741895],
-    title: "🍕",
-  },
-  { id: "marker-3", coordinate: [-118.3404, 38.3004029], title: "⭐️" },
-  { id: "marker-4", coordinate: [-118.3404, 38.4004029], title: "😇" },
-  { id: "marker-5", coordinate: [-118.3404, 38.3004029], title: "⭐️" },
-  { id: "marker-6", coordinate: [-122.4783, 37.7698646], title: "🌉" },
-  { id: "marker-7", coordinate: [-118.4975, 34.0094], title: "🎡" },
-  { id: "marker-8", coordinate: [-116.4975, 34.0094], title: "😐" },
-  { id: "marker-9", coordinate: [-118.4975, 34.0094], title: "😺" },
-  { id: "marker-10", coordinate: [2.3522, 48.8566], title: "🗼" },
-  { id: "marker-11", coordinate: [2.3509, 48.8566], title: "🏛" },
-  { id: "marker-12", coordinate: [7.4246, 43.7384], title: "🏖" },
-  { id: "marker-13", coordinate: [6.9807, 45.4408], title: "⛷" },
-  { id: "marker-14", coordinate: [-0.5792, 44.8374], title: "🍷" },
-];
-
-export const buttonsSelectData = [
-  { label: "Now" },
-  { label: "Today" },
-  { label: "Tomorrow" },
-  { label: "Next 7 Days" },
-  { label: "Next Week" },
-  { label: "Next 14 Days" },
-  { label: "Next Month" },
-  { label: "Next 30 Days" },
-  { label: "Custom" },
-];
